pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "changelog-ai"
version = "0.1.0"
description = "Generate CHANGELOG entries from git history using AI"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["git", "changelog", "ai", "llm", "developer-tools", "release-notes"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = [
    "click>=8.0",
    "gitpython>=3.1",
    "rich>=13.0",
    "anthropic>=0.25",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
changelog-ai = "changelog_ai.cli:main"

[project.urls]
Homepage = "https://github.com/faw21/changelog-ai"
Repository = "https://github.com/faw21/changelog-ai"

[tool.hatch.build.targets.wheel]
packages = ["src/changelog_ai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=changelog_ai --cov-report=term-missing"

[tool.coverage.run]
source = ["changelog_ai"]
